Add test for disk I/O usage
